edit

+++ Assignment1/assignment1_1_3/Program.cs
@@ -7,14 +7,14 @@ class Program {
     static void Main(string[] args) {
         //Ask user to input number. Use Parse function
         Console.Write("Please enter your first number (dividend): ");
-        int dividend = Convert.ToInt32(Console.ReadLine()); 
+        double dividend = Convert.ToDouble(Console.ReadLine()); 
 
         Console.Write("Please enter your second number (divisor): ");
-        int divisor = Convert.ToInt32(Console.ReadLine()); 
+        double divisor = Convert.ToDouble(Console.ReadLine()); 
 
         // Sum function
-        int quotient =  dividend / divisor;
-        int remainder = dividend % divisor;
+        double quotient =  dividend / divisor;
+        double remainder = dividend % divisor;
 
         //Print output
         Console.WriteLine($"Your total is: {quotient} with remainder: {remainder}");
